refactoring
moved login and token retrieval to credentials.npm
